+++ packages/rsocket-core/package.json
@@ -13,9 +13,7 @@
     "prepublishOnly": "yarn run build",
     "test": "yarn jest"
   },
-  "dependencies": {
-    "@rsocket/rsocket-types": "^1.0.0"
-  },
+  "dependencies": {},
   "devDependencies": {
     "rimraf": "~3.0.2",
     "typescript": "~4.3.0"
