+++ packages/rsocket-examples/package.json
@@ -14,7 +14,6 @@
   "dependencies": {
     "ws": "~8.0.0",
     "@rsocket/rsocket-core": "^1.0.0",
-    "@rsocket/rsocket-types": "^1.0.0",
     "@rsocket/rsocket-websocket-client": "^1.0.0"
   },
   "devDependencies": {
