+++ packages/rsocket-tcp-client/package.json
@@ -14,8 +14,7 @@
     "test": "jest"
   },
   "dependencies": {
-    "@rsocket/rsocket-core": "^1.0.0",
-    "@rsocket/rsocket-types": "^1.0.0"
+    "@rsocket/rsocket-core": "^1.0.0"
   },
   "devDependencies": {
     "rimraf": "~3.0.2",
