--- packages/rsocket-types/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@rsocket/rsocket-types",
-  "version": "1.0.0",
-  "main": "dist/index",
-  "types": "dist/index",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "yarn run clean && yarn run compile",
-    "clean": "rimraf -rf ./dist",
-    "compile": "tsc -p tsconfig.build.json",
-    "prepublishOnly": "yarn run build",
-    "test": "echo \"Error: no test specified\" && exit 0"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "rimraf": "~3.0.2",
-    "typescript": "~4.3.0"
-  }
-}
