+++ packages/rsocket-websocket-client/package.json
@@ -14,8 +14,7 @@
     "test": "echo \"Error: no test specified\" && exit 0"
   },
   "dependencies": {
-    "@rsocket/rsocket-core": "^1.0.0",
-    "@rsocket/rsocket-types": "^1.0.0"
+    "@rsocket/rsocket-core": "^1.0.0"
   },
   "devDependencies": {
     "rimraf": "~3.0.2",
