+++ packages/rsocket-websocket-client/src/WebsocketClientTransport.ts
@@ -1,4 +1,4 @@
-import { ClientTransport, DuplexConnection } from "@rsocket/rsocket-types";
+import { ClientTransport, DuplexConnection } from "@rsocket/rsocket-core";
 import { WebsocketDuplexConnection } from "./WebsocketDuplexConnection";
 
 export type ClientOptions = {
